Move account color picking to own class
