+++ app/k9mail/src/main/java/com/fsck/k9/account/AccountModule.kt
@@ -13,9 +13,17 @@ val newAccountModule = module {
         )
     }
 
+    factory {
+        AccountColorPicker(
+            preferences = get(),
+            resources = get(),
+        )
+    }
+
     factory<AccountSetupExternalContract.AccountCreator> {
         AccountCreator(
             accountCreatorHelper = get(),
+            accountColorPicker = get(),
             localFoldersCreator = get(),
             preferences = get(),
             context = androidApplication(),
