+++ app/ui/legacy/src/main/java/com/fsck/k9/account/AccountCreatorHelper.kt
@@ -1,10 +1,7 @@
 package com.fsck.k9.account
 
-import android.content.res.Resources
 import app.k9mail.core.common.mail.Protocols
 import com.fsck.k9.Account.DeletePolicy
-import com.fsck.k9.Preferences
-import com.fsck.k9.core.R
 import com.fsck.k9.mail.ConnectionSecurity
 
 /**
@@ -12,7 +9,7 @@ import com.fsck.k9.mail.ConnectionSecurity
  *
  * TODO move this close to account creator
  */
-class AccountCreatorHelper(private val preferences: Preferences, private val resources: Resources) {
+class AccountCreatorHelper {
 
     fun getDefaultDeletePolicy(type: String): DeletePolicy {
         return when (type) {
@@ -42,22 +39,5 @@ class AccountCreatorHelper(private val preferences: Preferences, private val res
 
     private fun getSmtpDefaultPort(connectionSecurity: ConnectionSecurity): Int {
         return if (connectionSecurity == ConnectionSecurity.SSL_TLS_REQUIRED) 465 else 587
-    }
-
-    fun pickColor(): Int {
-        val accounts = preferences.accounts
-        val usedAccountColors = accounts.map { it.chipColor }.toSet()
-        val accountColors = resources.getIntArray(R.array.account_colors).toList()
-
-        val availableColors = accountColors - usedAccountColors
-        if (availableColors.isEmpty()) {
-            return accountColors.random()
-        }
-
-        val defaultAccountColors = resources.getIntArray(R.array.default_account_colors)
-        return availableColors.shuffled().minByOrNull { color ->
-            val index = defaultAccountColors.indexOf(color)
-            if (index != -1) index else defaultAccountColors.size
-        } ?: error("availableColors must not be empty")
     }
 }
