+++ app/ui/legacy/src/main/java/com/fsck/k9/account/KoinModule.kt
@@ -14,7 +14,7 @@ val accountModule = module {
         )
     }
     factory { BackgroundAccountRemover(get()) }
-    factory { AccountCreatorHelper(get(), get()) }
+    factory { AccountCreatorHelper() }
     factory { (parameters: WorkerParameters) ->
         AccountRemoverWorker(accountRemover = get(), notificationController = get(), context = get(), parameters)
     }
